Move Planet classes to own package.
