+++ lib/src/main/kotlin/H3Writer.kt
@@ -1,4 +1,5 @@
 package com.eliottgray.kotlin
+import com.eliottgray.kotlin.planet.Planet
 import com.uber.h3core.H3Core
 import com.uber.h3core.LengthUnit
 import kotlinx.coroutines.*
