+++ lib/src/main/kotlin/planet/Planet.kt
@@ -1,5 +1,6 @@
-package com.eliottgray.kotlin
+package com.eliottgray.kotlin.planet
 
+import com.eliottgray.kotlin.*
 import com.github.benmanes.caffeine.cache.AsyncCache
 import com.github.benmanes.caffeine.cache.Caffeine
 import java.time.Duration
